Moved HueSlider and LightSaturationPreview into InputColor
