+++ packages/components/src/Form/Inputs/InputColor/HueSlider/index.ts
@@ -23,4 +23,5 @@
  SOFTWARE.
 
  */
-export * from './InputColorRevamp'
+
+export * from './HueSlider'
